+++ crates/swc/src/config/mod.rs
@@ -273,32 +273,6 @@ impl Options {
     }
 }
 
-/// Configuration related to source map generated by swc.
-#[derive(Clone, Serialize, Deserialize, Debug)]
-#[serde(untagged)]
-pub enum SourceMapsConfig {
-    Bool(bool),
-    Str(String),
-}
-
-impl SourceMapsConfig {
-    pub fn enabled(&self) -> bool {
-        match *self {
-            SourceMapsConfig::Bool(b) => b,
-            SourceMapsConfig::Str(ref s) => {
-                assert_eq!(s, "inline", "Source map must be true, false or inline");
-                true
-            }
-        }
-    }
-}
-
-impl Default for SourceMapsConfig {
-    fn default() -> Self {
-        SourceMapsConfig::Bool(true)
-    }
-}
-
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum InputSourceMap {
